# .static/swagger/swagger.yaml
definitions:
...
  api.loginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  api.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  api.registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  api.registerResponse:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
...
  api.tokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Returns access and refresh tokens.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.tokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Rotates refresh token and returns new access and refresh tokens.
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.tokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.registerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register user
      tags:
      - auth
...
swagger: "2.0"